requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.6
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
inflection==0.5.1
kombu==5.3.2
magic-filter==1.0.12
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
prompt-toolkit==3.0.39
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.6
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
yarl==1.9.2
